+++ NEWS.md
@@ -1,3 +1,17 @@
+# ukmaps 0.0.4
+
+* Fourth draft.
+* Relies solely on official data from May 2024.
+* The minimal level are Wards that aggregate to LADs, Counties, Regions, and Countries.
+* Uses the hierarchies from https://github.com/catmoez/UK-Local-Authority-districts.
+
+# ukmaps 0.0.3
+
+* Third draft.
+* Adds a function to plot the UK map at Local Authority Districts (LAD) (Local Government District (LGD) in
+    Northern Ireland)
+* Lighter final datasets.
+
 # ukmaps 0.0.2
 
 * Second draft. Provides columns to simplify grouping areas.
@@ -5,4 +19,3 @@
 # ukmaps 0.0.1
 
 * First draft.
-